requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
pydantic
python-dotenv
redis
celery
openai